+++ docs/changelog.rst
@@ -8,6 +8,7 @@ CHANGELOG
 **Bug fixes**
 
 - When updating interventions, stake field is no more required
+- Fix duplicates in year filters in intervention module
 
 
 2.24.5 (2019-03-06)