+++ docs/changelog.rst
@@ -12,6 +12,7 @@ CHANGELOG
 - Configurable blade code
 - Allow letters in blade number
 - Improve signage templates
+- Add "On signage/infrastructure" filter on intervention list
 
 
 2.24.5 (2019-03-06)